+++ src/main/java/com/niteroomcreation/simplewebrest/models/data/domains/Product.java
@@ -15,14 +15,14 @@ import javax.persistence.Id;
 @Entity
 public class Product {
 
-    private Integer id;
+    private Long id;
     private String name;
     private float price;
 
     public Product() {
     }
 
-    public Product(Integer id, String name, float price) {
+    public Product(Long id, String name, float price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,11 +30,11 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
